refactor: simplify test module structure per PR feedback
Moved tests from separate files to inline #[cfg(test)] module as suggested
by reviewer. This follows the more idiomatic Rust pattern for unit tests
and removes the non-standard src/tests/ directory structure.
